Hackily add new process mode, "grid"
This is mainly for my own purposes.

This takes in power-of-2-sized images and makes subimages for all 16x16
tiles in it. Each half of the image represents versions of tiles for the
left and right eye respectively at corresponding positions. The
subimages are therefore interleaved so that you can just add one to a
tile ID for the left subimage and get the one for the right. There's
also two dummy entries added at the front - this is for a quirk my own
engine that will probably change.